pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["motools", "mozoo"]

[project]
name = "motools"
version = "0.1.0"
description = "Infrastructure for training and evaluating model organisms"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "inspect-ai>=0.3.0",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.1.1",
    "datasets>=4.2.0",
    "typer>=0.19.2",
    "rich>=14.2.0",
    "scipy>=1.16.2",
    "matplotlib>=3.10.7",
    "tinker>=0.2.2",
    "transformers>=4.57.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Ignore line length in string literals/templates

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
exclude = ["tests/"]

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
    "mypy>=1.18.2",
    "pandas-stubs>=2.3.2.250926",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "respx>=0.22.0",
    "ruff>=0.14.0",
    "sphinx>=8.2.3",
    "sphinx-autodoc-typehints>=3.5.2",
    "sphinx-rtd-theme>=3.0.2",
    "types-aiofiles>=25.1.0.20251011",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
lint = "scripts.lint:main"
motools = "motools.cli.main:main"
